Symlinked the config files to HOME dir
Why?
 * To keep the config files and dotfiles in sync

+++ setup.sh
@@ -1,11 +1,20 @@
-cp .bash_aliases ~/
-cp .bash_functions ~/
-cp .git_aliases ~/
-cp .path_vars ~/
-cp .env_vars ~/
-cp .vimrc ~/
-cp todos.txt ~/
-cp quick_notes.txt ~/
-cp bucket_list.txt ~/
-cat .bashrc_partial >> ~/.bashrc
+# Link the config files to HOME directory counterparts
+ln -s $(pwd)/.bash_aliases $HOME/
+ln -s $(pwd)/.bash_functions $HOME/
+ln -s $(pwd)/.git_aliases $HOME/
+ln -s $(pwd)/.vimrc $HOME/
+[ -f ~/.path_vars ] || cp .path_vars ~/
+[ -f ~/.env_vars ] || cp .env_vars ~/
 
+# Copy note taking files
+[ -f ~/todos.txt ] || cp todos.txt ~/
+[ -f ~/quick_notes.txt ] || cp quick_notes.txt ~/
+[ -f ~/bucket_list.txt ] || cp bucket_list.txt ~/
+
+if [ "$1" == "mac" ]; then
+    cat .bash_cfg_partial >> ~/.bash_profile
+elif [ "$1" == "linux" ]; then
+    cat .bash_cfg_partial >> ~/.bashrc
+else
+    echo "Which os is this?"
+fi
